backend-python/requirements.txt
fastapi
uvicorn
python-dotenv
openai
httpx
